+++ src/access.rs
@@ -1,7 +1,48 @@
-use crate::{Access, AccessError, BitFlags, CompatError, CompatState, Compatibility, TryCompat};
+use crate::{
+    AccessError, AddRuleError, AddRulesError, BitFlags, CompatError, CompatState, Compatibility,
+    HandleAccessError, HandleAccessesError, Ruleset, TryCompat, ABI,
+};
+use enumflags2::BitFlag;
 
 #[cfg(test)]
 use crate::{make_bitflags, AccessFs};
+
+pub trait Access: PrivateAccess {
+    /// Gets the access rights defined by a specific [`ABI`].
+    /// Union of [`from_read()`](Access::from_read) and [`from_write()`](Access::from_write).
+    fn from_all(abi: ABI) -> BitFlags<Self> {
+        // An empty access-right would be an error if passed to the kernel, but because the kernel
+        // doesn't support Landlock, no Landlock syscall should be called.  try_compat() should
+        // also return RestrictionStatus::Unrestricted when called with unsupported/empty
+        // access-righs.
+        Self::from_read(abi) | Self::from_write(abi)
+    }
+
+    /// Gets the access rights identified as read-only according to a specific ABI.
+    /// Exclusive with [`from_write()`](Access::from_write).
+    fn from_read(abi: ABI) -> BitFlags<Self>;
+
+    /// Gets the access rights identified as write-only according to a specific ABI.
+    /// Exclusive with [`from_read()`](Access::from_read).
+    fn from_write(abi: ABI) -> BitFlags<Self>;
+}
+
+pub trait PrivateAccess: BitFlag {
+    fn ruleset_handle_access(
+        ruleset: &mut Ruleset,
+        access: BitFlags<Self>,
+    ) -> Result<(), HandleAccessesError>
+    where
+        Self: Access;
+
+    fn into_add_rules_error(error: AddRuleError<Self>) -> AddRulesError
+    where
+        Self: Access;
+
+    fn into_handle_accesses_error(error: HandleAccessError<Self>) -> HandleAccessesError
+    where
+        Self: Access;
+}
 
 // Creates an illegal/overflowed BitFlags<T> with all its bits toggled, including undefined ones.
 fn full_negation<T>(flags: BitFlags<T>) -> BitFlags<T>
