+++ src/ruleset.rs
@@ -1,9 +1,7 @@
 use crate::{
-    uapi, AccessFs, AddRuleError, AddRulesError, BitFlags, CompatState, Compatibility, Compatible,
-    CreateRulesetError, HandleAccessError, HandleAccessesError, RestrictSelfError, RulesetError,
-    TryCompat, ABI,
+    uapi, Access, AccessFs, AddRuleError, AddRulesError, BitFlags, CompatState, Compatibility,
+    Compatible, CreateRulesetError, RestrictSelfError, RulesetError, TryCompat, ABI,
 };
-use enumflags2::BitFlag;
 use libc::close;
 use std::io::Error;
 use std::mem::size_of_val;
@@ -11,43 +9,6 @@ use std::os::unix::io::RawFd;
 
 #[cfg(test)]
 use crate::*;
-
-pub trait Access: PrivateAccess {
-    /// Gets the access rights defined by a specific [`ABI`].
-    /// Union of [`from_read()`](Access::from_read) and [`from_write()`](Access::from_write).
-    fn from_all(abi: ABI) -> BitFlags<Self> {
-        // An empty access-right would be an error if passed to the kernel, but because the kernel
-        // doesn't support Landlock, no Landlock syscall should be called.  try_compat() should
-        // also return RestrictionStatus::Unrestricted when called with unsupported/empty
-        // access-righs.
-        Self::from_read(abi) | Self::from_write(abi)
-    }
-
-    /// Gets the access rights identified as read-only according to a specific ABI.
-    /// Exclusive with [`from_write()`](Access::from_write).
-    fn from_read(abi: ABI) -> BitFlags<Self>;
-
-    /// Gets the access rights identified as write-only according to a specific ABI.
-    /// Exclusive with [`from_read()`](Access::from_read).
-    fn from_write(abi: ABI) -> BitFlags<Self>;
-}
-
-pub trait PrivateAccess: BitFlag {
-    fn ruleset_handle_access(
-        ruleset: &mut Ruleset,
-        access: BitFlags<Self>,
-    ) -> Result<(), HandleAccessesError>
-    where
-        Self: Access;
-
-    fn into_add_rules_error(error: AddRuleError<Self>) -> AddRulesError
-    where
-        Self: Access;
-
-    fn into_handle_accesses_error(error: HandleAccessError<Self>) -> HandleAccessesError
-    where
-        Self: Access;
-}
 
 // Public interface without methods and which is impossible to implement outside this crate.
 pub trait Rule<T>: PrivateRule<T>
@@ -305,7 +266,7 @@ pub trait RulesetAttr: Sized + AsMut<Ruleset> {
     /// Consecutive calls to `handle_access()` will be interpreted as logical ORs
     /// with the previous handled accesses.
     ///
-    /// On error, returns a wrapped [`HandleAccessesError`].
+    /// On error, returns a wrapped [`HandleAccessesError`](crate::HandleAccessesError).
     /// E.g., `RulesetError::HandleAccesses(HandleAccessesError::Fs(HandleAccessError<AccessFs>))`
     fn handle_access<T, U>(mut self, access: T) -> Result<Self, RulesetError>
     where
